Library Functions Implemented
